Update machine learning notebook to use sqlalchemy + nlp.
The database connection and queries in machine learning.ipynb have been updated to use sqlalchemy instead of psycopg2.
This change was made to utilize the more robust and flexible features of sqlalchemy.
The ExecuteTime of each code cell is now recorded.
The data and models used in the notebook have also been updated to match the new database schema.
"Tweets", "temp_readings", and "fire_alerts" data is fetched and processed with the usage of pandas and tqdm libraries.
This creates a comprehensive dataframe for further machine learning model training.
Finally, an imbalance in training data is corrected using imblearn SMOTE and the processed data is saved for future use.